backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
sqlalchemy==2.0.36
gunicorn==23.0.0
dj-database-url==2.3.0
whitenoise==6.8.2